SME-35 refactor
