+++ pwa-shell/src/ui/menu/Drawer.js
@@ -4,7 +4,6 @@ import { Gesture } from 'react-with-gesture';
 
 import { ReactComponent as Close } from 'images/icons/ic_close.svg';
 import LogoutContainer from 'ui/menu/LogoutContainer';
-import KeycloakLogoutContainer from 'ui/menu/KeycloakLogoutContainer';
 
 class Drawer extends Component {
   constructor(props) {
@@ -53,12 +52,6 @@ class Drawer extends Component {
     const { isOpen } = this.props;
     const drawerState = isOpen ? 'open' : 'closed';
     const config = { event: { passive: false } };
-    const logoutContainer =
-      process.env.REACT_APP_AUTH_TYPE === 'keycloak' ? (
-        <KeycloakLogoutContainer />
-      ) : (
-        <LogoutContainer />
-      );
     return (
       <>
         <div className={`grey-overlay ${drawerState}`} />
@@ -81,7 +74,7 @@ class Drawer extends Component {
                     className="cursor-pointer color-primary-lightest float-right"
                   />
                 </div>
-                {logoutContainer}
+                <LogoutContainer />
                 {this.props.children}
               </div>
             );
