+++ pwa-shell/src/ui/menu/LogoutContainer.js
@@ -1,27 +1,12 @@
-import { connect } from 'react-redux';
-import { getUsername, logoutUser } from '@entando/apimanager';
+import React from 'react';
+import DefaultLogoutContainer from 'ui/menu/DefaultLogoutContainer';
+import KeycloakLogoutContainer from 'ui/menu/KeycloakLogoutContainer';
 
-import { isUserLogged } from 'state/content/selectors';
-import { fetchUserProfile } from 'state/thunks';
-import { getUserFullname } from 'state/user-profile/selectors';
-import { closeDrawer } from 'state/drawer/actions';
-import Logout from 'ui/menu/Logout';
+const LogoutContainer = () =>
+  process.env.REACT_APP_AUTH_TYPE === 'keycloak' ? (
+    <KeycloakLogoutContainer />
+  ) : (
+    <DefaultLogoutContainer />
+  );
 
-export const mapStateToProps = state => ({
-  isUserLogged: isUserLogged(state),
-  username: getUsername(state),
-  userFullname: getUserFullname(state),
-});
-
-export const mapDispatchToProps = dispatch => ({
-  loadUserProfile: username => dispatch(fetchUserProfile(username)),
-  logoutUser: () => {
-    dispatch(logoutUser());
-    dispatch(closeDrawer());
-  },
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(Logout);
+export default LogoutContainer;
